+++ frontend/src/components/@shared/SnackBar/type.ts
