+++ frontend/src/components/index.ts
@@ -18,6 +18,8 @@ export * from 'components/@shared/Tooltip';
 export * from 'components/@shared/ToggleButton';
 export * from 'components/@shared/UnderLineText';
 export * from 'components/@shared/Tabs';
+export * from 'components/@shared/SnackBar';
+export * from 'components/@shared/SROnly';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
 export * from 'components/Header';
@@ -30,7 +32,6 @@ export * from 'components/Timer';
 export * from 'components/ChallengeIcon';
 export * from 'components/Profile';
 export * from 'components/EmptyContent';
-export * from 'components/SnackBar';
 export * from 'components/PrivateOutlet';
 export * from 'components/LandingNavigation';
 export * from 'components/Title';
